Refactor Exporter to include apiPath and shared Collect method to reduce collectors boilerplate
